+++ src/query/functions/src/aggregates/aggregate_boolean.rs
@@ -27,11 +27,11 @@ use databend_common_expression::StateAddr;
 use databend_common_expression::StateSerdeItem;
 use databend_common_expression::SELECTIVITY_THRESHOLD;
 
-use super::aggregate_function_factory::AggregateFunctionDescription;
-use super::aggregate_function_factory::AggregateFunctionSortDesc;
 use super::assert_unary_arguments;
 use super::batch_merge1;
 use super::AggrState;
+use super::AggregateFunctionDescription;
+use super::AggregateFunctionSortDesc;
 use super::AggregateUnaryFunction;
 use super::FunctionData;
 use super::UnaryState;
@@ -186,7 +186,7 @@ pub fn try_create_aggregate_boolean_function<const IS_AND: bool>(
 }
 
 pub fn aggregate_boolean_function_desc<const IS_AND: bool>() -> AggregateFunctionDescription {
-    let features = super::aggregate_function_factory::AggregateFunctionFeatures {
+    let features = super::AggregateFunctionFeatures {
         is_decomposable: true,
         ..Default::default()
     };
