+++ src/query/functions/src/aggregates/aggregate_combinator_if.rs
@@ -28,15 +28,15 @@ use databend_common_expression::ProjectedBlock;
 use databend_common_expression::Scalar;
 use databend_common_expression::StateSerdeItem;
 
+use super::AggrState;
+use super::AggrStateLoc;
+use super::AggregateFunction;
+use super::AggregateFunctionCreator;
+use super::AggregateFunctionRef;
+use super::AggregateFunctionSortDesc;
+use super::CombinatorDescription;
 use super::StateAddr;
-use crate::aggregates::aggregate_function_factory::AggregateFunctionCreator;
-use crate::aggregates::aggregate_function_factory::AggregateFunctionSortDesc;
-use crate::aggregates::aggregate_function_factory::CombinatorDescription;
-use crate::aggregates::AggrState;
-use crate::aggregates::AggrStateLoc;
-use crate::aggregates::AggregateFunction;
-use crate::aggregates::AggregateFunctionRef;
-use crate::aggregates::StateAddrs;
+use super::StateAddrs;
 
 #[derive(Clone)]
 pub struct AggregateIfCombinator {
