+++ src/query/functions/src/aggregates/aggregate_kurtosis.rs
@@ -29,12 +29,12 @@ use databend_common_expression::StateAddr;
 use databend_common_expression::StateSerdeItem;
 use num_traits::AsPrimitive;
 
-use super::aggregate_function_factory::AggregateFunctionDescription;
-use super::aggregate_function_factory::AggregateFunctionSortDesc;
 use super::assert_unary_arguments;
 use super::batch_merge1;
 use super::AggrState;
+use super::AggregateFunctionDescription;
 use super::AggregateFunctionRef;
+use super::AggregateFunctionSortDesc;
 use super::AggregateUnaryFunction;
 use super::FunctionData;
 use super::UnaryState;
