+++ src/query/functions/src/aggregates/aggregate_markov_tarin.rs
@@ -44,13 +44,13 @@ use databend_common_expression::ProjectedBlock;
 use databend_common_expression::Scalar;
 use databend_common_expression::StateSerdeItem;
 
-use super::aggregate_function_factory::AggregateFunctionDescription;
 use super::assert_unary_arguments;
 use super::borsh_partial_deserialize;
 use super::extract_number_param;
+use super::AggrStateLoc;
 use super::AggregateFunction;
+use super::AggregateFunctionDescription;
 use super::StateAddr;
-use crate::aggregates::AggrStateLoc;
 
 pub struct MarkovTarin {
     display_name: String,
