+++ src/query/functions/src/aggregates/aggregate_quantile_tdigest.rs
@@ -39,17 +39,17 @@ use databend_common_expression::ScalarRef;
 use databend_common_expression::StateSerdeItem;
 use itertools::Itertools;
 
+use super::assert_params;
+use super::assert_unary_arguments;
 use super::borsh_partial_deserialize;
 use super::get_levels;
-use crate::aggregates::aggregate_function_factory::AggregateFunctionDescription;
-use crate::aggregates::aggregate_function_factory::AggregateFunctionSortDesc;
-use crate::aggregates::assert_params;
-use crate::aggregates::assert_unary_arguments;
-use crate::aggregates::AggrState;
-use crate::aggregates::AggrStateLoc;
-use crate::aggregates::AggregateFunction;
-use crate::aggregates::AggregateFunctionRef;
-use crate::aggregates::StateAddr;
+use super::AggrState;
+use super::AggrStateLoc;
+use super::AggregateFunction;
+use super::AggregateFunctionDescription;
+use super::AggregateFunctionRef;
+use super::AggregateFunctionSortDesc;
+use super::StateAddr;
 
 pub(crate) const MEDIAN: u8 = 0;
 pub(crate) const QUANTILE: u8 = 1;
