+++ src/query/functions/src/aggregates/aggregate_quantile_tdigest_weighted.rs
@@ -35,20 +35,20 @@ use databend_common_expression::Scalar;
 use databend_common_expression::StateSerdeItem;
 use num_traits::AsPrimitive;
 
+use super::aggregate_quantile_tdigest::QuantileTDigestState;
+use super::aggregate_quantile_tdigest::MEDIAN;
+use super::aggregate_quantile_tdigest::QUANTILE;
+use super::assert_binary_arguments;
+use super::assert_params;
 use super::borsh_partial_deserialize;
 use super::get_levels;
-use crate::aggregates::aggregate_function_factory::AggregateFunctionDescription;
-use crate::aggregates::aggregate_function_factory::AggregateFunctionSortDesc;
-use crate::aggregates::aggregate_quantile_tdigest::QuantileTDigestState;
-use crate::aggregates::aggregate_quantile_tdigest::MEDIAN;
-use crate::aggregates::aggregate_quantile_tdigest::QUANTILE;
-use crate::aggregates::assert_binary_arguments;
-use crate::aggregates::assert_params;
-use crate::aggregates::AggrState;
-use crate::aggregates::AggrStateLoc;
-use crate::aggregates::AggregateFunction;
-use crate::aggregates::AggregateFunctionRef;
-use crate::aggregates::StateAddr;
+use super::AggrState;
+use super::AggrStateLoc;
+use super::AggregateFunction;
+use super::AggregateFunctionDescription;
+use super::AggregateFunctionRef;
+use super::AggregateFunctionSortDesc;
+use super::StateAddr;
 
 #[derive(Clone)]
 pub struct AggregateQuantileTDigestWeightedFunction<T0, T1> {
