+++ src/query/functions/src/aggregates/aggregator.rs
@@ -15,6 +15,9 @@
 use super::aggregate_approx_count_distinct::aggregate_approx_count_distinct_function_desc;
 use super::aggregate_arg_min_max::aggregate_arg_max_function_desc;
 use super::aggregate_arg_min_max::aggregate_arg_min_function_desc;
+use super::aggregate_array_agg_function_desc;
+use super::aggregate_array_moving_avg_function_desc;
+use super::aggregate_array_moving_sum_function_desc;
 use super::aggregate_avg::aggregate_avg_function_desc;
 use super::aggregate_bitmap::aggregate_bitmap_and_count_function_desc;
 use super::aggregate_bitmap::aggregate_bitmap_intersect_count_function_desc;
@@ -23,43 +26,40 @@ use super::aggregate_bitmap::aggregate_bitmap_not_count_function_desc;
 use super::aggregate_bitmap::aggregate_bitmap_or_count_function_desc;
 use super::aggregate_bitmap::aggregate_bitmap_union_function_desc;
 use super::aggregate_bitmap::aggregate_bitmap_xor_count_function_desc;
+use super::aggregate_boolean::aggregate_boolean_function_desc;
 use super::aggregate_combinator_distinct::aggregate_combinator_distinct_desc;
 use super::aggregate_combinator_distinct::aggregate_combinator_uniq_desc;
 use super::aggregate_combinator_state::AggregateStateCombinator;
 use super::aggregate_covariance::aggregate_covariance_population_desc;
 use super::aggregate_covariance::aggregate_covariance_sample_desc;
+use super::aggregate_histogram_function_desc;
+use super::aggregate_json_array_agg_function_desc;
+use super::aggregate_json_object_agg_function_desc;
+use super::aggregate_kurtosis_function_desc;
 use super::aggregate_markov_tarin::aggregate_markov_train_function_desc;
+use super::aggregate_median_function_desc;
+use super::aggregate_median_tdigest_function_desc;
+use super::aggregate_median_tdigest_weighted_function_desc;
 use super::aggregate_min_max_any::aggregate_any_function_desc;
 use super::aggregate_min_max_any::aggregate_max_function_desc;
 use super::aggregate_min_max_any::aggregate_min_function_desc;
 use super::aggregate_mode::aggregate_mode_function_desc;
+use super::aggregate_quantile_cont_function_desc;
+use super::aggregate_quantile_disc_function_desc;
+use super::aggregate_quantile_tdigest_function_desc;
+use super::aggregate_quantile_tdigest_weighted_function_desc;
+use super::aggregate_range_bound_function_desc;
+use super::aggregate_retention_function_desc;
+use super::aggregate_skewness_function_desc;
+use super::aggregate_st_collect_function_desc;
 use super::aggregate_stddev::aggregate_stddev_pop_function_desc;
 use super::aggregate_stddev::aggregate_stddev_samp_function_desc;
+use super::aggregate_string_agg_function_desc;
+use super::aggregate_sum_function_desc;
 use super::aggregate_window_funnel::aggregate_window_funnel_function_desc;
 use super::AggregateCountFunction;
 use super::AggregateFunctionFactory;
 use super::AggregateIfCombinator;
-use crate::aggregates::aggregate_array_agg_function_desc;
-use crate::aggregates::aggregate_array_moving_avg_function_desc;
-use crate::aggregates::aggregate_array_moving_sum_function_desc;
-use crate::aggregates::aggregate_boolean::aggregate_boolean_function_desc;
-use crate::aggregates::aggregate_histogram_function_desc;
-use crate::aggregates::aggregate_json_array_agg_function_desc;
-use crate::aggregates::aggregate_json_object_agg_function_desc;
-use crate::aggregates::aggregate_kurtosis_function_desc;
-use crate::aggregates::aggregate_median_function_desc;
-use crate::aggregates::aggregate_median_tdigest_function_desc;
-use crate::aggregates::aggregate_median_tdigest_weighted_function_desc;
-use crate::aggregates::aggregate_quantile_cont_function_desc;
-use crate::aggregates::aggregate_quantile_disc_function_desc;
-use crate::aggregates::aggregate_quantile_tdigest_function_desc;
-use crate::aggregates::aggregate_quantile_tdigest_weighted_function_desc;
-use crate::aggregates::aggregate_range_bound_function_desc;
-use crate::aggregates::aggregate_retention_function_desc;
-use crate::aggregates::aggregate_skewness_function_desc;
-use crate::aggregates::aggregate_st_collect_function_desc;
-use crate::aggregates::aggregate_string_agg_function_desc;
-use crate::aggregates::aggregate_sum_function_desc;
 
 pub struct Aggregators;
