+++ src/query/functions/src/aggregates/mod.rs
@@ -64,8 +64,7 @@ pub use aggregate_combinator_if::AggregateIfCombinator;
 pub use aggregate_count::AggregateCountFunction;
 pub use aggregate_covariance::AggregateCovarianceFunction;
 pub use aggregate_function::*;
-pub use aggregate_function_factory::AggregateFunctionFactory;
-pub use aggregate_function_factory::AggregateFunctionSortDesc;
+pub use aggregate_function_factory::*;
 pub use aggregate_histogram::*;
 pub use aggregate_json_array_agg::*;
 pub use aggregate_json_object_agg::*;
